Cherry picked  from gatgui  fork : various  fixes related to  compile warnings  being removed.

+++ contrib/partio4Maya/partioEmitter.cpp
@@ -193,7 +193,7 @@ MStatus partioEmitter::initialize()
         eAttr.addField(formatExtMap[i].toUpperCase(),	i);
     }
 
-    eAttr.setDefault(Partio::readFormatIndex("pdc")); // PDC
+    eAttr.setDefault(short(Partio::readFormatIndex("pdc"))); // PDC
     eAttr.setChannelBox(true);
     eAttr.setKeyable(false);
