+++ contrib/partio4Maya/partioVisualizer.cpp
@@ -255,7 +255,7 @@ MStatus partioVisualizer::initialize()
         eAttr.addField(formatExtMap[i].toUpperCase(),	i);
     }
 
-    eAttr.setDefault(Partio::readFormatIndex("pdc"));  // PDC
+    eAttr.setDefault(short(Partio::readFormatIndex("pdc")));  // PDC
     eAttr.setChannelBox(true);
 
     aDrawStyle = eAttr.create( "drawStyle", "drwStyl");
@@ -777,15 +777,15 @@ MStatus partioVisualizer::compute( const MPlug& plug, MDataBlock& block )
         unsigned int numAttr=pvCache.particles->numAttributes();
         MPlug zPlug (thisMObject(), aPartioAttributes);
 
-        if ((colorFromIndex+1) > zPlug.numElements())
+        if ((colorFromIndex+1) > int(zPlug.numElements()))
         {
             block.outputValue(aColorFrom).setInt(-1);
         }
-        if ((opacityFromIndex+1) > zPlug.numElements())
+        if ((opacityFromIndex+1) > int(zPlug.numElements()))
         {
             block.outputValue(aAlphaFrom).setInt(-1);
         }
-        if ((radiusFromIndex+1) > zPlug.numElements())
+        if ((radiusFromIndex+1) > int(zPlug.numElements()))
         {
             block.outputValue(aRadiusFrom).setInt(-1);
         }
